create base resolver for web3
