+++ src/packages/auth/src/authentication/resolvers/web3/index.ts
@@ -1,1 +1,2 @@
+export * from './base-resolver';
 export * from './resolver';
